# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WOSP API Documentation",
        "description": "WOSP API in Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "WOSP API Server"
        }
    ],
    "paths": {
...
        "/collectors": {
            "get": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Get list of Collectors",
                "description": "Returns list of Collectors",
                "operationId": "getCollectorList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collectors"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Insert a new Collector",
                "description": "Insert a new Collector",
                "operationId": "postCollector",
                "requestBody": {
                    "description": "Collector to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CollectorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Collector created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collector"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Collector exist",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Istnieje już wolontariusz o podanym numerze w systemie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/collectors/{id}": {
            "get": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Get collector information",
                "description": "Returns collector data",
                "operationId": "getCollectorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Collector id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Collector"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/collectors/{collectorIdentifier}/box/create": {
            "post": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Create Charity Box for Collector",
                "description": "Create Charity Box for Collector",
                "operationId": "giveCollectorBox",
                "parameters": [
                    {
                        "name": "collectorIdentifier",
                        "in": "path",
                        "description": "Collector (volunteer) identifier id (number)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/collectors/{collectorIdentifier}/box/latestUncounted": {
            "get": {
                "tags": [
                    "Collectors"
                ],
                "summary": "Get last uncounted Collector's Charity Box",
                "description": "Get last uncounted Collector's Charity Box",
                "operationId": "getCollectorUncountedBox",
                "parameters": [
                    {
                        "name": "collectorIdentifier",
                        "in": "path",
                        "description": "Collector (volunteer) identifier id (number)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CharityBox"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Collectors": {
                "title": "Collectors",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Collector"
                        }
                    }
                },
                "type": "object"
            },
...
            "CollectorRequest": {
                "title": "CollectorRequest",
                "description": "new Collector schema",
                "required": [
                    "collectorIdentifier",
                    "firstName",
                    "lastName"
                ],
                "properties": {
                    "collectorIdentifier": {
                        "description": "Id of the collector identifier",
                        "format": "int64",
                        "example": "1"
                    },
                    "firstName": {
                        "description": "Wolunteer first name",
                        "type": "string",
                        "example": "Jan"
                    },
                    "lastName": {
                        "description": "Wolunteer last name",
                        "type": "string",
                        "example": "Kowalski"
                    }
                },
                "type": "object"
            },
...
            "CharityBox": {
                "title": "CharityBox",
                "description": "CharityBox model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Primary Key",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Time of created record",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Time of last record update",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "collectorIdentifier": {
                        "title": "Collector Identifier",
                        "description": "Identifier of volunteer's identification badge",
                        "type": "integer",
                        "format": "int64",
                        "example": 299
                    },
                    "collector": {
                        "$ref": "#/components/schemas/Collector"
                    },
                    "is_given_to_collector": {
                        "title": "Is Given To Collector",
                        "description": "Flag whether charity box has been issued",
                        "type": "boolean",
                        "example": 1
                    },
                    "given_to_collector_user_id": {
                        "$ref": "#/components/schemas/User"
                    },
                    "time_given": {
                        "title": "Time Given",
                        "description": "Time of giving charity box to volunter",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "is_counted": {
                        "title": "Is Counted",
                        "description": "Flag whether charity box were counted",
                        "type": "boolean",
                        "example": 0
                    },
                    "counting_user_id": {
                        "$ref": "#/components/schemas/User"
                    },
                    "time_counted": {
                        "title": "Time Counted",
                        "description": "Time of settlement charity box",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "is_confirmed": {
                        "title": "Is Confirmed",
                        "description": "Flag wheteher settlement of charity box were confirmed by other person",
                        "type": "boolean"
                    },
                    "user_confirmed_id": {
                        "title": "User Confirmed",
                        "description": "Id of user who confirmed charity box settlement",
                        "type": "integer"
                    },
                    "time_confirmed": {
                        "title": "TimeConfirmed",
                        "description": "Date confirmation of charity box settlement",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 9:57:30.196"
                    },
                    "count_1gr": {
                        "title": "Count_1gr",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 120
                    },
                    "count_2gr": {
                        "title": "Count_2gr",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 245
                    },
                    "count_5gr": {
                        "title": "Count_5gr",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 300
                    },
                    "count_10gr": {
                        "title": "Count_10gr",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 120
                    },
                    "count_20gr": {
                        "title": "Count_20gr",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 80
                    },
                    "count_50gr": {
                        "title": "Count_50gr",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_1zl": {
                        "title": "Count_1zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_2zl": {
                        "title": "Count_2zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_5zl": {
                        "title": "Count_5zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_10zl": {
                        "title": "Count_10zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_20zl": {
                        "title": "Count_20zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_50zl": {
                        "title": "Count_50zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_100zl": {
                        "title": "Count_100zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_200zl": {
                        "title": "Count_200zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "count_500zl": {
                        "title": "Count_500zl",
                        "description": "Number of coins",
                        "type": "int64",
                        "example": 60
                    },
                    "amount_PLN": {
                        "title": "Amount PLN",
                        "description": "Sum of collected money",
                        "type": "int64",
                        "example": 60
                    },
                    "amount_EUR": {
                        "title": "Amount EUR",
                        "description": "Sum of collected money",
                        "type": "int64",
                        "example": 60
                    },
                    "amount_USD": {
                        "title": "Amount USD",
                        "description": "Sum of collected money",
                        "type": "int64",
                        "example": 60
                    },
                    "amount_GBP": {
                        "title": "Amount GBP",
                        "description": "Sum of collected money",
                        "type": "int64",
                        "example": 60
                    },
                    "comment": {
                        "title": "Comment",
                        "description": "Comment",
                        "type": "string",
                        "example": "Lorem ipsum"
                    },
                    "is_special_box": {
                        "title": "Is Special Box",
                        "description": "Flag whether charity box is a special one",
                        "type": "boolean",
                        "example": 0
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CharityBox"
                }
            },
            "Collector": {
                "title": "Collector",
                "description": "Collector model - volunter with charity box",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Primary Key",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Time of created record",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Time of last record update",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "identifier": {
                        "title": "Identifier",
                        "description": "Volunteer ID number (number from the volunteer's badge)",
                        "type": "string",
                        "example": "123"
                    },
                    "first_name": {
                        "title": "First Name",
                        "description": "Volunteer's name",
                        "type": "string",
                        "example": "Jan"
                    },
                    "last_name": {
                        "title": "Last Name",
                        "description": "Volunteer's last name",
                        "type": "string",
                        "example": "Kowalski"
                    }
                },
                "type": "object"
            },
            "Role": {
                "title": "Role",
                "description": "User Role model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Primary Key",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Time of created record",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Time of last record update",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "name": {
                        "title": "Name",
                        "description": "User Role name",
                        "type": "string",
                        "enum": [
                            "volounteer",
                            "admin",
                            "superadmin"
                        ],
                        "example": "volunteer"
                    },
                    "description": {
                        "title": "Description",
                        "description": "User Role description",
                        "type": "string",
                        "example": "Wolontariusz liczący"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Primary Key",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "title": "Created At",
                        "description": "Time of created record",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "updated_at": {
                        "title": "Updated At",
                        "description": "Time of last record update",
                        "type": "string",
                        "format": "datetime",
                        "example": "2023-03-18 10:22:25.323"
                    },
                    "name": {
                        "title": "Name",
                        "description": "User's name",
                        "type": "string",
                        "example": "Jan"
                    },
                    "comment": {
                        "title": "Comment",
                        "description": "Comment",
                        "type": "string",
                        "example": "Lorem ipsum"
                    },
                    "roles": {
                        "title": "Roles",
                        "description": "User roles list",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "basic"
            }
        }
    },
...
}